flaskbackend/requirements.txt
Flask==2.3.2
Flask-PyMongo
Flask-Bcrypt
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
dnspython==2.3.0
requests==2.31.0
gunicorn
pytz
python-dateutil
APScheduler
